requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
pydantic==2.9.2
bcrypt==4.2.0
PyJWT==2.9.0
pytest==8.3.3
httpx==0.27.2
bandit==1.7.9
safety==3.2.8
